requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1
django-autoslug==1.9.8
django-crispy-forms==1.11.2
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
